+++ packages/x_storage_file/lib/src/file_storage_provider.dart
@@ -5,7 +5,7 @@ import 'package:dio/dio.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:x_storage_core/x_storage_core.dart';
 
-class FileXStorageDriver extends XStorageDriver with FileXStorageMixin {
+class FileStorageProvider extends XStorageProvider with FileProviderMixin {
   @override
   final String scheme = 'file';
 
@@ -13,7 +13,7 @@ class FileXStorageDriver extends XStorageDriver with FileXStorageMixin {
   final Dio _dio = Dio();
 
   @override
-  Future<void> deleteFile(XStorageUri uri) async {
+  Future<void> deleteFile(XUri uri) async {
     final filePath = await getFilePath(uri);
     final file = File(filePath);
 
@@ -23,7 +23,7 @@ class FileXStorageDriver extends XStorageDriver with FileXStorageMixin {
   }
 
   @override
-  Future<Uint8List?> loadFile(XStorageUri uri) async {
+  Future<Uint8List?> loadFile(XUri uri) async {
     final filePath = await getFilePath(uri);
     final file = File(filePath);
 
@@ -34,20 +34,21 @@ class FileXStorageDriver extends XStorageDriver with FileXStorageMixin {
   }
 
   @override
-  Future<void> saveFile(XStorageUri uri, Uint8List data) async {
+  Future<void> saveFile(XUri uri, Uint8List data) async {
     final filePath = await getFilePath(uri);
     final file = File(filePath);
     await file.writeAsBytes(data);
   }
 
   @override
-  Future<bool> exists(XStorageUri uri) async {
+  Future<bool> exists(XUri uri) async {
     final filePath = await getFilePath(uri);
     final file = File(filePath);
     return await file.exists();
   }
 
-  Future<String> getFilePath(XStorageUri uri) async {
+  @override
+  Future<String> getFilePath(XUri uri) async {
     final dirPath = await _getAppSupportDirPath();
     return '$dirPath${uri.path}';
   }
@@ -59,10 +60,10 @@ class FileXStorageDriver extends XStorageDriver with FileXStorageMixin {
   }
 }
 
-// /// Mixin for file system-based storage drivers
+// /// Mixin for file system-based storage providers
 // ///
 // /// Provides common functionality for handling file system storage.
-// mixin FileXStorageMixin on XStorageDriver {
+// mixin FileProviderMixin on XStorageProvider {
 //   @override
 //   XStorageType get storageType => XStorageType.file;
 // }
@@ -71,28 +72,29 @@ extension FileXStorageExtension on XStorage {
   /// ファイルをダウンロードして保存
   ///
   /// ダウンロードしたファイルのURIを返す
-  /// 対象のファイルのドライバーがNetworkXStorageMixinを実装している必要がある
+  /// 対象のファイルのプロバイダーがNetworkProviderMixinを実装している必要がある
   ///
-  Future<XStorageUri> downloadFile(
-    XStorageUri uri, {
+  Future<XUri> downloadFile(
+    XUri uri, {
     Function(int, int)? onProgress,
   }) async {
-    final driver = getDriver(uri);
-    final fileDriver = getDriverFromScheme<FileXStorageDriver>('file');
-    if (driver is! NetworkXStorageMixin) {
-      throw UnsupportedError('Unsupported driver type: ${driver.runtimeType}');
+    final provider = getProvider(uri);
+    final fileProvider = getProviderFromScheme<FileStorageProvider>('file');
+    if (provider is! NetworkProviderMixin) {
+      throw UnsupportedError(
+          'Unsupported provider type: ${provider.runtimeType}');
     }
 
-    final networkUrl = await driver.getNetworkUrl(uri);
-    final fileStorageUri = uri.changeScheme(fileDriver.scheme);
-    final filePath = await fileDriver.getFilePath(fileStorageUri);
+    final networkUrl = await provider.getNetworkUrl(uri);
+    final fileStorageUri = uri.changeScheme(fileProvider.scheme);
+    final filePath = await fileProvider.getFilePath(fileStorageUri);
 
     final file = File(filePath);
     if (await file.exists()) {
       return fileStorageUri;
     }
 
-    await fileDriver._dio.download(
+    await fileProvider._dio.download(
       networkUrl.toString(),
       filePath,
       onReceiveProgress: onProgress,