+++ packages/x_storage_file/lib/x_storage_file.dart
@@ -1,1 +1,1 @@
-export 'src/file_x_storage_driver.dart';
+export 'src/file_storage_provider.dart';