+++ packages/x_storage_firebase/lib/x_storage_firebase.dart
@@ -1,3 +1,3 @@
 library x_storage_firebase;
 
-export 'src/firebase_x_storage_driver.dart';
+export 'src/firebase_storage_provider.dart';