+++ packages/x_storage_presigned_url/lib/x_storage_presigned_url.dart
@@ -1,3 +1,3 @@
 library x_storage_presigned_url;
 
-export 'src/presigned_url_x_storage_driver.dart';
+export 'src/presigned_url_storage_provider.dart';