+++ packages/x_storage_core/CHANGELOG.md
@@ -1,3 +1,23 @@
+## 0.1.0
+
+* **BREAKING CHANGE**: Renamed `XStorageDriver` to `XStorageProvider` for better semantic clarity
+* Renamed all related classes, methods, and documentation to use "Provider" instead of "Driver"
+* Updated all mixins to work with `XStorageProvider`
+* Improved documentation
+
+## 0.0.4
+
+* Fix bugs
+
+## 0.0.3
+
+* Fix bugs
+
+## 0.0.2
+
+* Add getProviderFromScheme method to XStorage
+* Add changeSchema method to XUri
+
 ## 0.0.1
 
 * Initial release
@@ -5,7 +25,3 @@
 * Various Mixins (Network, File, Asset)
 * URI-based file management functionality
 
-## 0.0.2
-
-* Add getDriverFromScheme method to XStorage
-* Add changeSchema method to XStorageUri