+++ packages/x_storage_presigned_url/CHANGELOG.md
@@ -1,3 +1,22 @@
+## 0.1.0
+
+* **BREAKING CHANGE**: Renamed `PresignedUrlXStorageDriver` to `PresignedUrlXStorageProvider` for better semantic clarity
+* Updated to use x_storage_core 0.1.0
+* Updated http package to 1.2.0
+* Improved documentation
+
+## 0.0.4
+
+* Fix bugs
+
+## 0.0.3
+
+* Fix bugs
+
+## 0.0.2
+
+* Fix bugs
+
 ## 0.0.1
 
 * Initial release